[오승학] 결제 코드 추가, Sprite매니저 테스트 TopUI 테스트 유저 DB 로컬 테스트용 코드 추가

+++ Assets/Scripts/UI/Common/WealthUI.cs
@@ -9,8 +9,9 @@ public class WealthUI : MonoBehaviour
     [SerializeField] private TextMeshProUGUI wealthCountText;
     [SerializeField] private Image wealthImage;
 
-    public void SetWealth(string imageName, int wealthCount)
+    public void SetWealth(string imageName, long wealthCount)
     {
+        //Item가져다가 text설정하고, 샵으로 가는 것도 on off하도록 하게한다
         wealthCountText.text = wealthCount.ToString();
         wealthImage.sprite = SpriteManager.Instance.LoadSprite(imageName);
     }
